Backend file and dic-s renamed/refactored
