+++ src/main/java/com/codecool/tripplanner/DAO/NamedQueryHandler.java
@@ -1,4 +1,4 @@
-package com.codecool.tripplanner.searchHandler;
+package com.codecool.tripplanner.DAO;
 
 import com.codecool.tripplanner.JPA;
 import com.codecool.tripplanner.enums.CityName;
